+++ src/main/java/fun/wlfj/wyw1813004316jsp/dao/IBookDao.java
@@ -6,10 +6,10 @@ import fun.wlfj.wyw1813004316jsp.entity.Book;
 
 @Mapper
 public interface IBookDao {
-	public Book getBookById(int id);
+	public Book getBookById(long id);
 	public Book[] getBooks();
 	public void addBook(Book book);
-	public void deleteBook(int id);
+	public void deleteBook(long id);
 	public void updateBook(Book book);
 	public Book[] getBookByKeyword(String keyword);
 }
